requirements.txt
asgiref==3.2.10
awesome-slugify==1.6.5
Django==3.1.1
django-environ==0.4.5
django-widget-tweaks==1.4.8
flake8==3.8.3
gunicorn==20.0.4
importlib-metadata==1.7.0
mccabe==0.6.1
Pillow==7.2.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
regex==2020.7.14
reportlab==3.5.49
sqlparse==0.3.1
Unidecode==0.4.21
zipp==3.1.0
